LPS-70113 No need to use impl inside internal - forms-and-workflow
