+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/render/BaseNumberDDMFormFieldValueRenderer.java
@@ -12,28 +12,23 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.internal.render.impl;
+package com.liferay.dynamic.data.mapping.internal.render;
 
-import com.liferay.dynamic.data.mapping.model.DDMFormFieldType;
 import com.liferay.dynamic.data.mapping.model.Value;
+import com.liferay.dynamic.data.mapping.render.BaseDDMFormFieldValueRenderer;
 import com.liferay.dynamic.data.mapping.render.ValueAccessor;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.util.HtmlUtil;
-import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.GetterUtil;
+
+import java.text.NumberFormat;
 
 import java.util.Locale;
 
 /**
  * @author Marcellus Tavares
  */
-public class TextHTMLDDMFormFieldValueRenderer
-	extends BaseTextDDMFormFieldValueRenderer {
-
-	@Override
-	public String getSupportedDDMFormFieldType() {
-		return DDMFormFieldType.TEXT_HTML;
-	}
+public abstract class BaseNumberDDMFormFieldValueRenderer
+	extends BaseDDMFormFieldValueRenderer {
 
 	@Override
 	protected ValueAccessor getValueAcessor(Locale locale) {
@@ -43,21 +38,17 @@ public class TextHTMLDDMFormFieldValueRenderer
 			public String get(DDMFormFieldValue ddmFormFieldValue) {
 				Value value = ddmFormFieldValue.getValue();
 
-				return StringUtil.replace(
-					_HTML,
-					new String[] {"[$DDM_FORM_FIELD_VALUE$]", "[$PREVIEW$]"},
-					new String[] {
-						HtmlUtil.escapeJS(value.getString(locale)),
-						LanguageUtil.get(locale, "preview")
-					});
+				String valueString = value.getString(locale);
+
+				Number number = GetterUtil.getNumber(valueString);
+
+				NumberFormat numberFormat = NumberFormat.getNumberInstance(
+					locale);
+
+				return numberFormat.format(number);
 			}
 
 		};
 	}
-
-	private static final String _HTML =
-		"<a href=\"javascript:;\" onclick=\"Liferay.DDLUtil." +
-			"openPreviewDialog('[$DDM_FORM_FIELD_VALUE$]');\">([$PREVIEW$])" +
-				"</a>";
 
 }
