+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/render/IntegerDDMFormFieldValueRenderer.java
@@ -12,19 +12,19 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.internal.render.impl;
+package com.liferay.dynamic.data.mapping.internal.render;
 
 import com.liferay.dynamic.data.mapping.model.DDMFormFieldType;
 
 /**
  * @author Marcellus Tavares
  */
-public class DecimalDDMFormFieldValueRenderer
+public class IntegerDDMFormFieldValueRenderer
 	extends BaseNumberDDMFormFieldValueRenderer {
 
 	@Override
 	public String getSupportedDDMFormFieldType() {
-		return DDMFormFieldType.DECIMAL;
+		return DDMFormFieldType.INTEGER;
 	}
 
 }
