+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/render/TextAreaDDMFormFieldValueRenderer.java
@@ -12,19 +12,19 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.internal.render.impl;
+package com.liferay.dynamic.data.mapping.internal.render;
 
 import com.liferay.dynamic.data.mapping.model.DDMFormFieldType;
 
 /**
  * @author Marcellus Tavares
  */
-public class RadioDDMFormFieldValueRenderer
-	extends BaseListDDMFormFieldValueRenderer {
+public class TextAreaDDMFormFieldValueRenderer
+	extends BaseTextDDMFormFieldValueRenderer {
 
 	@Override
 	public String getSupportedDDMFormFieldType() {
-		return DDMFormFieldType.RADIO;
+		return DDMFormFieldType.TEXT_AREA;
 	}
 
 }
